Add uterm_display_ready()
Split uterm_display_bind() in two steps.
In the atomic drm use case, all displays of a video card must be
setup together. So first enable all displays and then call the seat
and terminal callback to make use of the available displays.

The reason is that a GPU might be able to handle a 4K display on DP1
or on DP2, but not both at the same time. So the displays can't be
setup independantly.

Signed-off-by: Jocelyn Falempe <jfalempe@redhat.com>